ASTF dynamic profile is available in DP side. (#9)
* Support multiple instances for profiles
* Support duration for each profile
* Support RX packet handling for each profile
* Adapt the concept of updating global tuneables
* Clear thread statistics for compatible operation
* Keeping profile statistics after it stopped
* Stop profile gracefully: wait 2 sec

+++ src/astf/astf_db.cpp
@@ -35,8 +35,8 @@ inline std::string methodName(const std::string& prettyFunction)
 
 #define __METHOD_NAME__ methodName(__PRETTY_FUNCTION__)
 
-// make the class singelton
-CAstfDB* CAstfDB::m_pInstance = NULL;
+// make the class singleton
+std::unordered_map<uint32_t, CAstfDB*> CAstfDB::m_pInstances;
 
 
 CAstfDB::CAstfDB(){
@@ -57,6 +57,10 @@ CAstfDB::~CAstfDB(){
         delete m_validator;
     }
     delete m_topo_mngr;
+    for (auto it : m_smart_gen) {
+        delete it.second;
+    }
+    m_smart_gen.clear();
 }
 
 
@@ -130,6 +134,19 @@ void CTcpDataAssocTranslation::clear() {
     }
     m_map.clear();
     m_vec.clear();
+}
+
+void CTcpDataAssocTranslation::enumerate_server_ports(std::vector<uint16_t>& ports, bool is_stream) {
+    for (CTcpDataAssocTransHelp& trans_help : m_vec) {
+        if (trans_help.m_params.m_stream == is_stream) {
+            ports.push_back(trans_help.m_params.m_port);
+        }
+    }
+    for (assoc_map_it_t it = m_map.begin(); it != m_map.end(); it++) {
+        if (it->first.m_stream == is_stream) {
+            ports.push_back(it->first.m_port);
+        }
+    }
 }
 
 bool CAstfDB::start_profile_no_buffer(Json::Value msg){
@@ -897,10 +914,9 @@ bool CAstfDB::read_tunables(CTcpTuneables *tune, Json::Value tune_json) {
     return true;
 }
 
-ClientCfgDB  *CAstfDB::get_client_db() {
-    static ClientCfgDB g_dummy;
+ClientCfgDB  *CAstfDB::get_client_cfg_db() {
     if (m_client_config_info==0) {
-        return  &g_dummy;
+        return  &m_client_config_db;
     }else{
         return m_client_config_info;
     }
@@ -1016,6 +1032,9 @@ CAstfTemplatesRW *CAstfDB::get_db_template_rw(uint8_t socket_id, CTupleGenerator
     // json data should not be accessed by multiple threads in parallel
     std::unique_lock<std::mutex> my_lock(m_global_mtx);
 
+    if (g_gen == nullptr) {
+        g_gen = get_smart_gen(thread_id);
+    }
     g_gen->Create(0, thread_id);
 
     uint16_t rss_thread_id, rss_thread_max;
@@ -1066,7 +1085,7 @@ CAstfTemplatesRW *CAstfDB::get_db_template_rw(uint8_t socket_id, CTupleGenerator
         if (ip_gen_list[i]["dir"] == "c") {
             gen_idx_trans.push_back(last_c_idx);
             last_c_idx++;
-            ClientCfgDB  * cdb=get_client_db();
+            ClientCfgDB  * cdb=get_client_cfg_db();
             g_gen->add_client_pool(dist, portion.m_ip_start, portion.m_ip_end, active_flows_per_core, *cdb, 0, 0);
         } else {
             gen_idx_trans.push_back(last_s_idx);
@@ -1125,8 +1144,7 @@ CAstfTemplatesRW *CAstfDB::get_db_template_rw(uint8_t socket_id, CTupleGenerator
         CTcpTuneables *s_tuneable;
         CTcpTuneables *c_tuneable = new CTcpTuneables();
         assert(c_tuneable);
-        assert(CAstfDB::m_pInstance);
-        s_tuneable = CAstfDB::m_pInstance->get_s_tune(index);
+        s_tuneable = get_s_tune(index);
         assert(s_tuneable);
 
         if (!read_tunables(c_tuneable, m_val["templates"][index]["client_template"]["glob_info"])){
@@ -1435,27 +1453,35 @@ const std::vector<std::string>& CAstfDB::get_tg_names() {
     return m_tg_names;
 }
 
-void CAstfDB::clear_db_ro_rw(CTupleGeneratorSmart *g_gen) {
+void CAstfDB::clear_db_ro_rw(CTupleGeneratorSmart *g_gen, uint16_t thread_id) {
     std::unique_lock<std::mutex> my_lock(m_global_mtx);
-    for ( auto &rw_db : m_rw_db) {
-        if ( !rw_db ) {
-            continue;
+    if (g_gen == nullptr) {
+        remove_smart_gen(thread_id);
+    }
+    else {
+        g_gen->Delete();
+    }
+    /* to clear only when there is no thread generator */
+    if (m_smart_gen.size() == 0) {
+        for ( auto &rw_db : m_rw_db) {
+            if ( !rw_db ) {
+                continue;
+            }
+            rw_db->Delete();
+            delete rw_db;
         }
-        rw_db->Delete();
-        delete rw_db;
-    }
-    m_rw_db.clear();
-    for (auto &s_tuneables : m_s_tuneables) {
-        delete s_tuneables;
-    }
-    m_s_tuneables.clear();
-    m_prog_lens.clear();
+        m_rw_db.clear();
+        for (auto &s_tuneables : m_s_tuneables) {
+            delete s_tuneables;
+        }
+        m_s_tuneables.clear();
+        m_prog_lens.clear();
 
-    for (auto &tcp_data : m_tcp_data) {
-        tcp_data.Delete();
+        for (auto &tcp_data : m_tcp_data) {
+            tcp_data.Delete();
+        }
+        m_json_initiated = false;
     }
-    g_gen->Delete();
-    m_json_initiated = false;
     my_lock.unlock();
 }
